Cleanup: rename utils.cpp, flatten detail/ headers, rationalize names, update CHANGELOG
- Rename finufft_utils.cpp → utils.cpp; update CMakeLists.txt, makefile, setupIWYU.cmake
- Flatten detail/ subdirectory: merge all headers into include/finufft/
- Fix stale header references; add spread_1d/2d/3d.cpp TUs
- IWYU: trim stale includes, add missing standard headers
- Rationalize internal headers: 13 → 10 files with shorter names
- Update CHANGELOG; reformat in-source changelog comments
