+++ include/finufft/makeplan.hpp
@@ -1,15 +1,16 @@
 #pragma once
 
+#include <algorithm>
 #include <cmath>
+#include <complex>
 #include <cstdio>
-#include <memory>
+#include <limits>
+#include <type_traits>
 #include <vector>
 
-#include <finufft/finufft_core.hpp>
-#include <finufft/finufft_utils.hpp>
-#include <finufft/heuristics.hpp>
+#include <finufft/plan.hpp>
+#include <finufft/utils.hpp>
 #include <finufft/spreadinterp.hpp>
-#include <finufft/xsimd.hpp>
 #include <finufft_common/kernel.h>
 #include <finufft_common/utils.h>
 
@@ -20,7 +21,7 @@ int FINUFFT_PLAN_T<TF>::set_nf_type12(BIGINT ms, BIGINT *nf) const
 // Type 1 & 2 recipe for how to set 1d size of upsampled array, nf, given opts
 // and requested number of Fourier modes ms. Returns 0 if success, else an
 // error code if nf was unreasonably big (& tell the world).
-// 2/24/26 Barbone: converted from free function to method on FINUFFT_PLAN_T.
+// Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
 // Previous args (opts, spopts) are now plan members; only ms and nf remain.
 {
   using namespace finufft::utils;
@@ -68,7 +69,7 @@ void FINUFFT_PLAN_T<TF>::onedim_fseries_kernel(BIGINT nf,
   Fixed num_threads 7/20/20. Reduced rounding error in a[n] calc 8/20/24.
   11/25/25, replaced kernel_definition by evaluate_kernel_runtime, meaning that
   the FT of the piecewise poly approximant (not "exact" kernel) is computed.
-  2/24/26 Barbone: converted from free function to method on FINUFFT_PLAN_T.
+  Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
   Previous arg opts (spreading opts) is now read from plan member spopts.
  */
 {
