+++ include/finufft/simd.hpp
@@ -1,17 +1,83 @@
 #pragma once
 
-#include <finufft/finufft_utils.hpp>
-#include <finufft/spreadinterp.hpp>
+#include <finufft/plan.hpp>
+#include <finufft/utils.hpp>
 #include <finufft_common/kernel.h>
 
-#include <finufft/xsimd.hpp>
+// xsimd configuration: ensure unsupported architectures fall back to emulated.
+#include <xsimd/config/xsimd_config.hpp>
+#ifdef XSIMD_NO_SUPPORTED_ARCHITECTURE
+#undef XSIMD_NO_SUPPORTED_ARCHITECTURE
+#define XSIMD_WITH_EMULATED 1
+#define XSIMD_DEFAULT_ARCH  emulated<128>
+#endif
+#include <xsimd/xsimd.hpp>
 
 #include <array>
 #include <cinttypes>
 #include <cmath>
+#include <cstddef>
 #include <cstdio>
 #include <cstdlib>
+#include <type_traits>
 #include <vector>
+
+// SIMD width utilities (moved from utils.hpp to keep utils.hpp SIMD-free).
+#if defined(XSIMD_VERSION_MAJOR) && (XSIMD_VERSION_MAJOR >= 14)
+namespace finufft::utils {
+
+template<class T, uint8_t N = 1> constexpr uint8_t min_simd_width() {
+  // finds the smallest simd width that can handle N elements
+  // simd size is batch size the SIMD width in xsimd terminology
+  if constexpr (std::is_void_v<xsimd::make_sized_batch_t<T, N>>) {
+    return min_simd_width<T, N * 2>();
+  } else {
+    return N;
+  }
+};
+
+template<class T, uint8_t N> constexpr std::size_t find_optimal_simd_width() {
+  // finds the smallest simd width that minimizes the number of iterations
+  // NOTE: might be suboptimal for some cases 2^N+1 for example
+  // in the future we might want to implement a more sophisticated algorithm
+
+  uint8_t optimal_simd_width = min_simd_width<T>();
+  uint8_t min_iterations     = (N + optimal_simd_width - 1) / optimal_simd_width;
+  for (uint8_t simd_width = optimal_simd_width;
+       simd_width <= xsimd::batch<T, xsimd::best_arch>::size; simd_width *= 2) {
+    uint8_t iterations = (N + simd_width - 1) / simd_width;
+    if (iterations < min_iterations) {
+      min_iterations     = iterations;
+      optimal_simd_width = simd_width;
+    }
+  }
+  return static_cast<std::size_t>(optimal_simd_width);
+}
+
+template<class T, uint8_t N> constexpr std::size_t GetPaddedSIMDWidth() {
+  // helper function to get the SIMD width with padding for the given number of elements
+  // that minimizes the number of iterations
+
+  return xsimd::make_sized_batch<T, find_optimal_simd_width<T, N>()>::type::size;
+}
+template<class T, uint8_t ns>
+constexpr std::size_t get_simd_width_helper(uint8_t runtime_ns) {
+  if constexpr (ns < finufft::common::MIN_NSPREAD) {
+    return static_cast<std::size_t>(0);
+  } else {
+    if (runtime_ns == ns) {
+      return GetPaddedSIMDWidth<T, ns>();
+    } else {
+      return get_simd_width_helper<T, ns - 1>(runtime_ns);
+    }
+  }
+}
+template<class T> constexpr std::size_t GetPaddedSIMDWidth(int runtime_ns) {
+  return get_simd_width_helper<T, 2 * ::finufft::common::MAX_NSPREAD>(runtime_ns);
+}
+
+} // namespace finufft::utils
+#endif // XSIMD_VERSION_MAJOR >=14
 
 namespace finufft::spreadinterp {
