+++ include/finufft/spread.hpp
@@ -1,13 +1,13 @@
 #pragma once
 
-#include <finufft/detail/simd_helpers.hpp>
+#include <finufft/simd.hpp>
 
 namespace finufft::spreadinterp {
 } // namespace finufft::spreadinterp
 
 // ---------- FINUFFT_PLAN_T spread_subproblem_*_kernel method definitions ----------
 // FINUFFT_PLAN_T is already defined via the transitive include chain:
-//   detail/simd_helpers.hpp -> finufft/spreadinterp.hpp -> finufft/finufft_core.hpp
+//   simd.hpp -> finufft/plan.hpp
 
 template<typename TF>
 template<int ns, int nc>
@@ -33,7 +33,7 @@ FINUFFT_NEVER_INLINE void FINUFFT_PLAN_T<TF>::spread_subproblem_1d_kernel(
      chance of segfault when epsmach*N1>O(1), assuming max() and ceil() commute.
      This needed off1 as extra arg. AHB 11/30/20.
      Vectorized using xsimd by M. Barbone 06/24.
-     2/24/26 Barbone: converted from free function template to method on FINUFFT_PLAN_T.
+     Converted from free function template to method on FINUFFT_PLAN_T. Barbone 2/24/26.
      Previous arg horner_coeffs_ptr is now read from plan member horner_coeffs.data().
   */
   using namespace finufft::spreadinterp;
@@ -169,7 +169,7 @@ FINUFFT_NEVER_INLINE void FINUFFT_PLAN_T<TF>::spread_subproblem_2d_kernel(
    dd (size M complex) are complex source strengths
    du (size size1*size2) is complex uniform output array
    For algoritmic details see spread_subproblem_1d_kernel.
-   2/24/26 Barbone: converted from free function template to method on FINUFFT_PLAN_T.
+   Converted from free function template to method on FINUFFT_PLAN_T. Barbone 2/24/26.
    Previous arg horner_coeffs_ptr is now read from plan member horner_coeffs.data().
 */
 {
@@ -267,8 +267,8 @@ FINUFFT_NEVER_INLINE void FINUFFT_PLAN_T<TF>::spread_subproblem_3d_kernel(
     const UBIGINT size2, const UBIGINT size3, TF *FINUFFT_RESTRICT du, const UBIGINT M,
     const TF *kx, const TF *ky, const TF *kz, const TF *dd) const noexcept
 // 3D version of spread_subproblem_1d_kernel.
-// 2/24/26 Barbone: converted from free function template to method on FINUFFT_PLAN_T.
-// Previous arg horner_coeffs_ptr is now read from plan member horner_coeffs.data().
+// Converted from free function template to method on FINUFFT_PLAN_T.
+// Previous arg horner_coeffs_ptr is now read from plan member horner_coeffs.data(). Barbone 2/24/26.
 {
   using namespace finufft::spreadinterp;
   using finufft::common::MAX_NSPREAD;
@@ -360,7 +360,8 @@ void FINUFFT_PLAN_T<TF>::add_wrapped_subgrid(
    using atomic writes (R Blackwell, Nov 2020).
    Merged the thread_safe and the not thread_safe version of the function into one
    (M. Barbone 06/24).
-   Accesses plan members nfdim[0..2] for grid dimensions N1,N2,N3.
+   Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
+   Previous args (N1, N2, N3) are now read from plan member nfdim[0..2].
 */
 {
   using T          = TF;
@@ -436,6 +437,10 @@ void FINUFFT_PLAN_T<TF>::bin_sort_singlethread(double bin_size_x, double bin_siz
  *         writes to ret a vector list of indices, each in the range 0,..,M-1.
  *         Thus, ret must have been preallocated for M BIGINTs.
  *
+ * Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
+ * Previous args (ret, M, kx, ky, kz, N1, N2, N3, debug) are now plan members
+ * (sortIndices, nj, XYZ[0..2], nfdim[0..2], spopts.debug).
+ *
  * Notes: I compared RAM usage against declaring an internal vector and passing
  * back; the latter used more RAM and was slower.
  * Avoided the bins array, as in JFM's spreader of 2016,
@@ -505,7 +510,9 @@ void FINUFFT_PLAN_T<TF>::bin_sort_multithread(double bin_size_x, double bin_size
    Explicit #threads control argument 7/20/20.
    Improved by Martin Reinecke, 6/19/23 (up to 50% faster at 1 thr/core).
    Todo: if debug, print timing breakdowns.
-   Accesses plan members: sortIndices, nj, XYZ, nfdim.
+   Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
+   Previous args (ret, M, kx, ky, kz, N1, N2, N3, debug) are now plan members
+   (sortIndices, nj, XYZ[0..2], nfdim[0..2], spopts.debug).
  */
 {
   using namespace finufft::spreadinterp;
@@ -621,7 +628,8 @@ void FINUFFT_PLAN_T<TF>::get_subgrid(BIGINT &offset1, BIGINT &offset2, BIGINT &o
    inaccurate) single-precision with N1>>1e7 on 11/30/20.
    3) Requires O(M) RAM reads to find the k array bnds. Almost negligible in
    tests.
-   Accesses plan members: spopts.nspread, dim.
+   Converted from free function to method on FINUFFT_PLAN_T. Barbone 2/24/26.
+   Previous args (ns, ndims) are now read from plan members spopts.nspread and dim.
 */
 {
   using namespace finufft::spreadinterp;
@@ -656,7 +664,7 @@ void FINUFFT_PLAN_T<TF>::get_subgrid(BIGINT &offset1, BIGINT &offset2, BIGINT &o
 }
 
 // ---------- FINUFFT_PLAN_T spread-subproblem nested caller definitions ----------
-// Out-of-class definitions of the nested types declared in finufft_core.hpp.
+// Out-of-class definitions of the nested types declared in plan.hpp.
 // Member function templates are not allowed in local classes (GCC restriction),
 // so these must be proper nested class definitions of FINUFFT_PLAN_T<TF>.
 
@@ -727,15 +735,15 @@ struct FINUFFT_PLAN_T<TF>::SpreadSubproblem3dCaller {
 
 // ---------- FINUFFT_PLAN_T spread-subproblem method definitions ----------
 // FINUFFT_PLAN_T is already defined via the transitive include chain:
-//   detail/simd_helpers.hpp -> finufft/spreadinterp.hpp -> finufft/finufft_core.hpp
+//   simd.hpp -> finufft/plan.hpp
 
 template<typename TF>
 void FINUFFT_PLAN_T<TF>::spread_subproblem_1d(BIGINT off1, UBIGINT size1, TF *du,
                                               UBIGINT M, TF *kx, TF *dd) const noexcept
 // Spread M NU points (kx, dd) into subgrid du of length size1 starting at off1.
 // Uses plan members spopts.nspread, nc, horner_coeffs for the kernel dispatch.
-// 2/24/26 Barbone: converted from free function to method on FINUFFT_PLAN_T.
-// Previous args (opts, horner_coeffs_ptr, nc) are now plan members.
+// Converted from free function to method on FINUFFT_PLAN_T.
+// Previous args (opts, horner_coeffs_ptr, nc) are now plan members. Barbone 2/24/26.
 {
   using namespace finufft::spreadinterp;
   using namespace finufft::common;
@@ -752,8 +760,8 @@ void FINUFFT_PLAN_T<TF>::spread_subproblem_2d(BIGINT off1, BIGINT off2, UBIGINT 
                                               UBIGINT M, const TF *kx, const TF *ky,
                                               const TF *dd) const noexcept
 // 2D version of spread_subproblem_1d.
-// 2/24/26 Barbone: converted from free function to method on FINUFFT_PLAN_T.
-// Previous args (opts, horner_coeffs_ptr, nc) are now plan members.
+// Converted from free function to method on FINUFFT_PLAN_T.
+// Previous args (opts, horner_coeffs_ptr, nc) are now plan members. Barbone 2/24/26.
 {
   using namespace finufft::spreadinterp;
   using namespace finufft::common;
@@ -770,8 +778,8 @@ void FINUFFT_PLAN_T<TF>::spread_subproblem_3d(BIGINT off1, BIGINT off2, BIGINT o
                                               TF *du, UBIGINT M, TF *kx, TF *ky, TF *kz,
                                               TF *dd) const noexcept
 // 3D version of spread_subproblem_1d.
-// 2/24/26 Barbone: converted from free function to method on FINUFFT_PLAN_T.
-// Previous args (opts, horner_coeffs_ptr, nc) are now plan members.
+// Converted from free function to method on FINUFFT_PLAN_T.
+// Previous args (opts, horner_coeffs_ptr, nc) are now plan members. Barbone 2/24/26.
 {
   using namespace finufft::spreadinterp;
   using namespace finufft::common;
