+++ include/finufft_common/common.h
@@ -4,4 +4,5 @@
 #include <finufft_common/defines.h>
 #include <finufft_common/kernel.h>
 #include <finufft_common/pswf.h>
+#include <finufft_common/safe_call.h>
 #include <finufft_common/utils.h>
