+++ perftest/spreadtestndall.cpp
@@ -10,7 +10,7 @@
 
  Todo: Add some more args + update screen output like spreadtestnd.cpp.
  */
-#include "finufft/finufft_utils.hpp"
+#include "finufft/utils.hpp"
 #include <finufft/test_defs.hpp>
 
 #include <cmath>
