+++ src/makeplan.cpp
@@ -1,4 +1,4 @@
-#include <finufft/detail/makeplan.hpp>
+#include <finufft/makeplan.hpp>
 
 // Explicit instantiations, selected by FINUFFT_SINGLE define.
 // Private members are included because they are called from setpts
@@ -18,7 +18,7 @@ template int FINUFFT_PLAN_T<FLT>::setup_spreadinterp();
 template void FINUFFT_PLAN_T<FLT>::precompute_horner_coeffs();
 
 // These are called by init_grid_kerFT_FFT (instantiated in fft.cpp), so must be
-// explicitly instantiated here where their bodies (in detail/makeplan.hpp) are visible.
+// explicitly instantiated here where their bodies (in makeplan.hpp) are visible.
 template int FINUFFT_PLAN_T<FLT>::set_nf_type12(BIGINT ms, BIGINT *nf) const;
 template void FINUFFT_PLAN_T<FLT>::onedim_fseries_kernel(BIGINT nf,
                                                           std::vector<FLT> &fwkerhalf) const;
