+++ src/spread_2d.cpp
@@ -1,0 +1,13 @@
+#include <finufft/spread.hpp>
+
+// Explicit instantiation of 2D spread subproblem, selected by FINUFFT_SINGLE define.
+
+#ifdef FINUFFT_SINGLE
+using FLT = float;
+#else
+using FLT = double;
+#endif
+
+template void FINUFFT_PLAN_T<FLT>::spread_subproblem_2d(
+    BIGINT, BIGINT, UBIGINT, UBIGINT, FLT *FINUFFT_RESTRICT, UBIGINT, const FLT *,
+    const FLT *, const FLT *) const noexcept;
