+++ test/finufft1dmany_test.cpp
@@ -1,6 +1,6 @@
 #include <finufft/test_defs.hpp>
 // this enforces recompilation, responding to SINGLE...
-#include "finufft/finufft_utils.hpp"
+#include "finufft/utils.hpp"
 #include "utils/dirft1d.hpp"
 #include "utils/norms.hpp"
 using namespace std;
